功能增强：新增国家战略关系属性系统
核心功能：
- 新增战略关系数据库模型，支持五级关系（战争/冲突/无外交/伙伴/盟友）
- 自动初始化大国/超级大国之间的战略关系
- 在决策引擎中集成战略关系信息，LLM决策时可查看各国关系
- 提供战略关系管理API端点

数据库变更：
- 新增strategic_relationship表存储双边战略关系
- 更新simulation_project模型添加关系关联

后端变更：
- 新增StrategicRelationshipEnum服务
- 更新AgentInfo和InfoPool支持战略关系
- 仿真服务集成关系初始化和数据加载
- API路由添加战略关系查询和设置接口

前端变更：
- 优化仿真结果页面展示
- 更新统计API和数据结构

+++ app/core/agent_base.py
@@ -37,8 +37,8 @@ class ActionConfig(BaseModel):
     action_category: str
     action_desc: str
     respect_sov: bool
-    initiator_power_change: int
-    target_power_change: int
+    initiator_power_change: float
+    target_power_change: float
     is_initiative: bool
     is_response: bool
